Fixes #129 - Changes "About Us" into a dropdown menu for "Who We Are" and "Leadership" (#182)
Not 100% sure if we need this to open when we hover, or if just clicking is sufficient. The built in behavior for the hovering isn't that great, becuase there is a gap between the menu and the dropdown - so either we need to artificially move the menu up when it appears (like it does with the language dropdown), or else have some kind of debounce value for whether or not the menu should be open. Both are solvable issues, but probably don't need to solve it now.